apikey update

+++ cypress/support/step_definitions/api_admin_remove_api_key.js
@@ -4,23 +4,50 @@
 //    When the API admin user removes registry admin API keys from the entity registry
 //    Then the API keys no longer provide access to the entity registry
 
+function createTestEntity(apikey, expectedStatus) {
+	cy.fixture('entityTestData.json')
+	.then((testData) => {
+		let createEntityUrl = "/registry/" + registryName + '/entity';
+		cy.request({
+			url: createEntityUrl,
+			method: 'POST',
+			failOnStatusCode: false,
+			headers: {
+				'x-api-key': apiKey
+			}
+		}).then((response) => {
+			assert.equals(response.status, expectedStatus)
+		})
+	})
+
+}
 
 when('the API admin user removes registry admin API keys from the entity registry', () =>{
 	cy.get("@registryName").then((registryName) => {
-		cy.get("@apiAdminApiKey").then((apiKey) => {
+		cy.get("@apiAdminApiKey").then((apiAdminApiKey) => {
 			let addApikeyUrl = "/registry/" + registryName + "/apikey";
 			cy.fixture('newApiKeyData.json')
 			.then((testData) => {
+				cy.wrap(testData.apikey).as('newApiKey');
 				cy.request({
 					url: addApikeyUrl,
-					method: 'DELETE',
+					method: 'POST',
 					headers: {
-						'apiKey': apiKey
+						'x-api-key': apiAdminApiKey
 					},
 					body: testData
-				}).then((response) => {
-					// test return from create
-					cy.log(response.body)
+				})
+				cy.get('@newApiKey').then((newApiKey) => {
+					createTestEntity(newApiKey, 200); // test that there is a valid apikey that provide access to the entity registry
+				})
+				
+				let deleteApiKeyUrl = "/registry/" + registryName + "/apikey/" + 
+				cy.request({
+					url: deleteApiKeyUrl,
+					method: 'DELETE',
+					headers: {
+						'x-api-key': apiAdminApiKey
+					}
 				})
 			})
 		})
@@ -29,8 +56,8 @@ when('the API admin user removes registry admin API keys from the entity registr
 
 then('the API keys no longer provide access to the entity registry', () => {
 	cy.get('@returnId').then((returnId) => {
-		let getEntityUrl = "/registry/" + registryName + "/entity/" + returnId;
-
-		cy.request(getEntityUrl)
+		cy.get('@newApiKey').then((newApiKey) => {
+			createTestEntity(newApiKey, 403);
+		})
 	})
 })
